requirements.txt
asgiref==3.5.0
autopep8==1.6.0
boto3==1.21.13
botocore==1.24.13
cycler==0.11.0
Django==4.0.3
fonttools==4.31.2
jmespath==0.10.0
kiwisolver==1.4.2
mat4py==0.5.0
matplotlib==3.5.1
mysql-connector-python==8.0.28
numpy==1.22.3
packaging==21.3
Pillow==9.0.1
protobuf==3.19.4
pycodestyle==2.8.0
pyparsing==3.0.7
python-dateutil==2.8.2
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tk==0.1.0
toml==0.10.2
urllib3==1.26.8
